+++ Models/Dto/CourseRegistrationViewDto.cs
@@ -8,7 +8,7 @@
         public string? CourseCategory { get; set; }
         public decimal? CourseFee { get; set; }
         public DateTime RegistrationDate { get; set; }
-        public int CourseScheduleId { get; set; }
-        public int Id { get; set; }
+        public long CourseScheduleId { get; set; }
+        public long Id { get; set; }
     }
 }
