+++ Services/Interfaces/IAcademics.cs
@@ -6,9 +6,9 @@ namespace EduReg.Services.Interfaces
     public interface IAcademics
     {
         Task<GeneralResponse> CreateAcademicLevel(AcademicLevelsDto model);
-        Task<GeneralResponse> UpdateAcademicLevelAsync(int Id, AcademicLevelsDto model);
-        Task<GeneralResponse> DeleteAcademicLevelAsync(int Id);
-        Task<GeneralResponse> GetAcademicLevelByIdAsync(int Id);
+        Task<GeneralResponse> UpdateAcademicLevelAsync(long Id, AcademicLevelsDto model);
+        Task<GeneralResponse> DeleteAcademicLevelAsync(long Id);
+        Task<GeneralResponse> GetAcademicLevelByIdAsync(long Id);
         Task<GeneralResponse> GetAllAcademicLevelAsync();
     }
 }
