+++ Services/Interfaces/IFaculties.cs
@@ -6,9 +6,9 @@ namespace EduReg.Services.Interfaces
     public interface IFaculties
     {
         Task<GeneralResponse> CreateFacultyAsync(FacultiesDto model);
-        Task<GeneralResponse> UpdateFacultyAsync(int Id, FacultiesDto model);
-        Task<GeneralResponse> DeleteFacultyAsync(int Id);
-        Task<GeneralResponse> GetFacultyByIdAsync(int Id);
+        Task<GeneralResponse> UpdateFacultyAsync(long Id, FacultiesDto model);
+        Task<GeneralResponse> DeleteFacultyAsync(long Id);
+        Task<GeneralResponse> GetFacultyByIdAsync(long Id);
         Task<GeneralResponse> GetFacultyByCodeAsync(string facultyCode);
         Task<GeneralResponse> GetAllFacultiesAsync();
     }
